refactor: fileUpload 상태관리를 훅으로 분리
